+++ algebra/number/util.py
@@ -1,7 +1,8 @@
 import collections
-import heapq
 import math
 from typing import Dict, List
+
+from algebra.util.queue import HeapQueueSet
 
 
 def is_prime(num: int) -> bool:
@@ -48,25 +49,6 @@ def divisor_function(number):
     for p, e in factorize(number).items():
         n *= e + 1
     return n
-
-
-class HeapQueueSet:
-    def __init__(self):
-        self.queue = []
-        self.queue_set = set()
-
-    def size(self):
-        return len(self.queue)
-
-    def push(self, i):
-        if i not in self.queue_set:
-            heapq.heappush(self.queue, i)
-            self.queue_set.add(i)
-
-    def pop(self):
-        i = heapq.heappop(self.queue)
-        self.queue_set.remove(i)
-        return i
 
 
 def divisor_list(number) -> List[int]:
